Moved the tests of FailableMemoryAllocator to TestMemoryAllocatorTest file
