+++ Makefile.am
@@ -161,8 +161,7 @@ CppUTestTests_SOURCES = \
 	tests/TestResultTest.cpp \
 	tests/TestUTestMacro.cpp \
 	tests/UtestTest.cpp \
-	tests/UtestPlatformTest.cpp\
-	tests/FailableMemoryAllocatorTest.cpp
+	tests/UtestPlatformTest.cpp
 
 CppUTestExtTests_CPPFLAGS = $(lib_libCppUTestExt_a_CPPFLAGS)
 CppUTestExtTests_CFLAGS = $(lib_libCppUTestExt_a_CFLAGS)
