+++ tests/CMakeLists.txt
@@ -31,7 +31,6 @@ set(CppUTestTests_src
     UtestTest.cpp
     SimpleMutexTest.cpp
     UtestPlatformTest.cpp
-    FailableMemoryAllocatorTest.cpp
 )
 
 if (MSVC)
